requirements.txt
Django==5.0.7
djangorestframework==3.15.2
celery==5.4.0
redis==5.0.7
django-celery-beat==2.6.0
django-celery-results==2.5.1
graphene_django==3.2.2
graphene==3.3